output completed - no errors

+++ src/vector.c
@@ -31,7 +31,7 @@ Vector *vector_init(int size, destroy_vector destroy_fn, compare_vector compare_
 void vector_push_back(Vector *v, data_vector data)
 {
     if (v->size >= v->capacity)
-    {   
+    {
         v->capacity *= 2;
         v->data = (data_vector *)realloc(v->data, v->capacity * sizeof(data_vector));
     }
@@ -70,9 +70,12 @@ data_vector vector_get(Vector *v, int idx)
 
 void vector_destroy(Vector *v)
 {
-    for (int i = 0; i < v->size; i++)
+    if (v->destroy_fn)
     {
-        v->destroy_fn(v->data[i]);
+        for (int i = 0; i < v->size; i++)
+        {
+            v->destroy_fn(v->data[i]);
+        }
     }
     free(v->data);
     free(v);
